Переносит функцию расчета звания из мока
